+++ lib/xqsr3/extensions/test/unit/assert_not.rb
@@ -6,9 +6,9 @@ module Assertions
 
 	unless respond_to? :assert_not
 
-		def assert_not(test, failure_message = '')
+		def assert_not(expression, failure_message = '')
 
-			assert !(test), failure_message
+			assert !(expression), failure_message
 		end
 	end