+++ CHANGES.md
@@ -1,5 +1,15 @@
 # **xqsr3** Changes
 
+## 0.32.2 - 12th April 2019
+
+* ~ substantially improved performance of Enumerable#unique()
+* ~ fixes to FrequencyMap
+* ~ fixes to MultiMap
+
+## 0.32.1 - 12th April 2019
+
+* FrequencyMap : ~ @counts => @elements
+
 ## 0.32.0 - 12th April 2019
 
 * ~ refactored Xqsr3::Containers::FrequencyMap, and corrected #push() to handle -ve count